Authenticate using simple JWT token
only checking if token is on header, and if so, using key to verify it.
Additinally gettign the user from local db, and checking the expirity
date as well. This way can still use default django login methods and
decorators

Fall back as default django auth (session) for dev purposes if
gatekeeper url is not defined, otherwise login will redirect to it.
this should solve most of #2 but still need to test out with gatekeeper
to make sure