initial version of Activtiy Type web ui management
using a slightly different visualisation, might change in the future.
For now just want to test it out not having multiple concerns on the
same django view. In the future we might want to separate like this as
well, or if we really need to we can try to refactor the
list+update+create view to be simpler or more modularised.