setting up to use agstack Asset Registry if config is set
also requires AGSTACK_CLIENT_SECRET and AGSTACK_API_KEY, otherwise if
not set it will simply fall back to the UUID5 with the geometry string.

also added simple optimisation to only send the request to agstack AR if the geometry of the parcel is new. (new parcel or changed the geometry of an existing parcel).

tested with a local AR, using the global Agstack Usermanager. So should
be ok to use with the global AR and have the same results, but still
need to test it out. refs #11