enforncing a nested activity form with correct field exclusion
only inherithing the base class with general field exclusion for the activity forms was giving inconsistent behavious in the child form classes. To avoid this we need to ensure that the child form classes don't have any required field in the field exclusion variable (`exclude`) in the meta class.

I saw this inconsistency in the behaviour of the irrigation operation (as a nested activty of a compost operation), which doesn't work. Mean while the same form works for the AddRawMaterialOperation as a nested operation for compost operation. With this fix, both cases works correctly.